+++ widget/src/lib.rs
@@ -27,6 +27,7 @@ mod row;
 
 pub mod button;
 pub mod checkbox;
+pub mod combo_box;
 pub mod container;
 pub mod overlay;
 pub mod pane_grid;
@@ -62,6 +63,8 @@ pub use button::Button;
 pub use checkbox::Checkbox;
 #[doc(no_inline)]
 pub use column::Column;
+#[doc(no_inline)]
+pub use combo_box::ComboBox;
 #[doc(no_inline)]
 pub use container::Container;
 #[doc(no_inline)]